Remove AbstractAlgoPreparation
Besides sharing two lines of codes the preparations have nothing in common and we never use the two implementing classes in an abstract way anyway. Also doWork does not allow us to use different return types for the different preparations.
